requirements.txt
pydantic>=1.10
pydantic<3
python-dotenv>=1.0
python-dotenv<2
requests>=2.31
requests<3
beautifulsoup4>=4.12
beautifulsoup4<5
instructor>=1.11.0
openai>=1.30
openai<2
pandas>=2.0
pandas<3
pytest>=7.0
pytest<8
pytest-asyncio>=0.21
pytest-asyncio<0.25
google-genai>=1.31.0
fastapi-cli>=0.0.8
fastapi-cloud-cli>=0.1.4
httpx>=0.28.1
httpx<1.0.0
typer>=0.15.1
psutil>=7
google-cloud-bigquery>=3.14
google-cloud-bigquery<4
tqdm>=4.66
tqdm<5
playwright>=1.44
